gnu: Remove Go 1.9.
We can wait to push this patch until there is another reason to change
the Go compiler, but I am submitting it now in case anyone else is
working on Go in Guix.

* gnu/packages/golang.scm (go-1.9): Remove variable.
(go-1.11): Inherit from go-1.4 and consolidate the inherited build
phases from go-1.9.

Go 1.11 should be essentially unchanged with this commit, although the
derivation is different.